.bucket_versions.yaml file stays in bucket all times
Instead of having a local copy of the file that tracks the bucket version, just keep in the s3 bucket.

Whenever a command is run, get latest version from the bucket.
If modification/creation command, then the version in the bucket is updated.

Hopefully reduce chances of version conflicts/collisions.