Refactor tsv_query to layered architecture
In the CLI layer, tsv_query checks if the tsv exists on disk. If not,
it used modules from file_cli.py to check if it is in the bucket. If so,
it downloads the tsv file. The tsv is then passed to the query layer,
attemps a query, returns the query string and results for printing
in the CLI layer.