Configure job system testing environment
The default behaviour of multiple compose files is to append the port
mappings, which will lead to conflicts when attempting to spin up the
orignal stack and the testing stack at the same time, as both will try
to use the same ports. The !override tag ensures that the port value
is replaced instead of appended, allowing both stacks to run
concurrently.