Make pytest spin up testing env for job system
The override compose.yaml changes host ports to avoid conflicts with the
main job system stack, if that happens to also be running on the same
machine. Tempfs is used for the redis and flower logs, which allows for
testing of logs during the test session while still purging them after
the session ends.

Another key addition is a helper function that ensures that the docker
compose health checks are successful before job system tests are run.