Include MinIO in the docker compose stack
In the main stack, mount a volume that can persist for MinIO data. To
avoid user permission issues, use root user for the MinIO container for
now when doing local dev. For testing stack, use tmpfs  and override
ports to allow both stacks to run at the same time without port
conflicts (just like was used for the other containers).