Add fastAPI container to docker-compose
Built with the goal to make divbase-cli jobs status work. (divbase-cli
jobs submit will fail since it does not support downloading of files
from minio yet).

For the fastAPI container, use the host 0.0.0.0 for the uvicorn
server to ensure it is accessible from outside the container. Might want
to refine this in divbase_api.py later to reduce duplicatation/overides.

For fastAPI to connect to the flower container in the stack, refer to it
with the env var http://flower:5555 instead of localhost:5555. Keep
localhost as a fallback in task_history.py so ensure that divbase-cli
query task-status still work (does not go via fastAPI).

Add a health endpoint to API that returns {"status": "ok"} for docker
compose healthcheck.