Improve docker volume garbage collection
To avoid anonymous volumes being created and risk hanging around on disk
as garbage, add volume mounts for rabbitmq and minio, based on the how
the anonymous mounts are used when not specified. This info can be found
in the mounts section of the logs from docker inspect <container_id> for
the containers running rabbitmq and minio, respectively.