Add a second celery worker and queue for long jobs
Some findings: there is a default queue called 'celery' that is used by
Celery if no queue is specified. The 'quick' worker is set to handle
tasks from both the 'quick' and 'celery' queues, while the 'long'
worker is set to handle tasks from the 'long' queue.

The concurrency of the workers is set to 4 to accomodate that there are
now two workers on the host machine that need to share the available CPU
resources. If not specified, the concurrency will be automatically set
based on the CPUs of the host machine, but this is not adjusted across
workers so it might lead to issues with too high loads if submitting
many jobs at once like in the pytest. The concurrency value for the
workers will later need to be adjusted for deployment on the cluster.