Use dynamic routing instead of static routing
Replace static routing with dynamic routing. For now, just use simple
pattern matching of task names to decide routing. In the future, routing
could be more advanced, e.g. based on the size of a VCF file to be
queried with bcftools.

Update the test_task_routing test so that it can handle both static
routing and dynamic routing for its assertions.