Use bcftools_pipe_task in the concurrency test
This allows us to drop simulate_quick_task from tasks.py (which is good
since we don't want that in production). The downside is that the test
could be flaky on some machines since it is dependent on the timing of
running the task. Current wait time is 0.01 seconds and was chosen based
on previous runtime in the flower log. 0.001 s was too short on my
machine.