chown the app contents for compose watch
From docs:
watch also requires that the container's USER can write to the target path so it can update files. A common pattern is for initial content to be copied into the container using the COPY instruction in a Dockerfile. To ensure such files are owned by the configured user, use the COPY --chown flag:

https://docs.docker.com/compose/how-tos/file-watch/