Add script for running a demo divbase workflow
A mini-workflow with the currently required steps to run a query in
divbase using a medium-to-large sized datasset.

Ensures: that a 5.5 Gb public VCF is downloaded, that mock metadata is
generated, that the files are put in a bucket in the local minIO
container, that a query job that takes about 10 minutes is queued, and
that the dimensions (samples and variants ) of the VCF file are
calculated and logged.