Add unique identifier to bcftools pipe files
To avoid collisions when running concurrent bcftools queries. Each query
will now append the celery task-id to the temp and merged output files.

Later on, it is likely that the output files will use another naming
convention, but for now task-id is enough to avoid collisions between
jobs.

If there is no celery task-id (when running task directly without the
queue), use the current timestamp as a unique identifier. This might not
be robust for cases when multiple tasks are started at the same time,
but all tasks in prod will use the job system so it is unlikely to be an
issue.