Add scripts for benchmarking with factorial design
The intention of these scripts is to analyze of how the dimensions of
VCF files (samples, variants) affect the runtime of bcftools queries in
divbase. The scripts are a draft and may require further refinement, but
can already be used to run a factorial design experiment on the divbase
architechture.