Skip transfer of unused VCFs for 'view -r' jobs
When running queries that include `bcftools view -r`, the system will
now check the dimension index file to see if there are files that do not
contain the regions that will be filtered on. If so, these files will
not be transferred from the bucket to the worker. This will potentially
save a lot of resources when the bucket contains large VCF files.

Example:
Init a clean compose stack

docker compose -f docker/divbase_compose.yaml down -v && \
docker compose -f docker/divbase_compose.yaml up --build -d; \
python scripts/local_dev_setup.py

if this query is run when the bucket does not contain
.vcf_dimensions.yaml, it will transfer all VCF found by the metadata tsv
query to the worker:

DIVBASE_ENV=local divbase-cli query bcftools-pipe
--tsv-filter "Area:West of Ireland;Sex:F" \
--command "view -s SAMPLES; view -r 1,4,6,21,24" \
--metadata-tsv-name sample_metadata_HOM_chr_split_version.tsv \
--project local-project-3

But if this command has been run before to create the dimensions file,
the task will identify potential VCFs from the metadata tsv query that
do not contain the regions that will be filtered on,
and skip transferring them:

DIVBASE_ENV=local divbase-cli dimensions update \
--project local-project-3