Use bcftools concat instead of merge when needed
From close inspection of the output file of the bcftools query command,
it appears that bcftools merge is not supposed to be used for cases when
the input files (here: the temp files) have overlapping sample names.
The reason it worked was because --force-samples was used. The bcftools
manual says to use concat when sample names overlap. This commit adds
initial logic to check the temp files for when to use merge or concat.