Support all bcftools view -r cases in early checks
According to the bcftools view manual, the syntax is for view -r is:
"chr|chr:pos|chr:from-to|chr:from". It was discovered during the work on
the early exit logic that this was not fully implemented in the check
for which files to transfer from the bucket to the worker.

DIVBASE_ENV=local divbase-cli dimensions update \
--project local-project-1 && \
DIVBASE_ENV=local divbase-cli query bcftools-pipe \
--tsv-filter "Area:Northern Portugal" --command "view -s SAMPLES; \
view -r 1,3,4:200,21:15000000-25000000" \
--metadata-tsv-name sample_metadata.tsv --project local-project-1