Add dimensions upload fail error to task log
Using bcftools_pipe_task as precedent, errors in tasks should preferably
be logged to the task log to ensure that users can see the error
message. Errors on upload of the dimensions file to S3 is now
propagated to the task level where it is caught and written to the task
log.