Use multi-stage builds and drop -e for pip install
This almost cuts the image sizes in half. The key was to drop editable
mode from pip install, since that requires that both the source code
and the installed packages are in the image. Reduced code duplication is
good. But it seems that hot-reloading with Docker Compose watch needs
pip install -e, so that might be lost.