move VCFDimensionsEntryMissingError to api layer
The cli's make_authenticated_request already raises DivBaseAPIError for api exceptions
so plugging in VCFDimensionsEntryMissingError into that mechanism.

Tasks.py also updated to raise ValueError instead of that specific error, with TODO - to change to return dict with status error, as I think that is desired way for the worker failures?