Use signal handlers to get timestamps for runtime
Task enqueued signal handler (after_task_publish) and the tasks in
worker signal handlers (calling on _upsert_task_timestamps) are now
designed so that the handle different fields of the entry, and all
signal handlers are allowed to INSERT the entry to the TaskHistoryDB
table in case of race conditions. This should ensure that TaskHistoryDB
entries will have correct fields regardless of the order that signal
handlers happen to fire.